build: Move type marshalling library from 'marshal' to 'src/tss2-mu'.
Minor fixups were required to a few 'include' statements in a few test
cases. Also this commit fixes up the pkgconfig file for libtss2-mu.
