+++ Makefile.am
@@ -27,7 +27,7 @@
 #;**********************************************************************;
 
 ### Initialize global variables used throughout the file ###
-INCLUDE_DIRS    = -I$(srcdir) -I$(srcdir)/include/tss2
+INCLUDE_DIRS    = -I$(srcdir) -I$(srcdir)/src -I$(srcdir)/include/tss2
 ACLOCAL_AMFLAGS = -I m4
 AM_CFLAGS       = $(INCLUDE_DIRS) $(EXTRA_CFLAGS) $(CODE_COVERAGE_CFLAGS)
 AM_LDFLAGS      = $(EXTRA_LDFLAGS) $(CODE_COVERAGE_LIBS)
@@ -97,17 +97,17 @@ libutil_la_CFLAGS = $(AM_CFLAGS)
 libutil_la_SOURCES = log/log.c log/log.h tcti/tcti.c tcti/tcti.h
 
 ### TCG TSS Marshaling/Unmarshalling spec library ###
-libmarshal = marshal/libmarshal.la
+libtss2_mu = src/tss2-mu/libtss2-mu.la
 tss2_HEADERS += $(srcdir)/include/tss2/tss2_mu.h
-lib_LTLIBRARIES += $(libmarshal)
-nodist_pkgconfig_DATA += lib/marshal.pc
-EXTRA_DIST += lib/libmarshal.map lib/marshal.pc.in
+lib_LTLIBRARIES += $(libtss2_mu)
+nodist_pkgconfig_DATA += lib/tss2-mu.pc
+EXTRA_DIST += lib/tss2-mu.map lib/tss2-mu.pc.in
 
 if HAVE_LD_VERSION_SCRIPT
-marshal_libmarshal_la_LDFLAGS = -Wl,--version-script=$(srcdir)/lib/libmarshal.map
+src_tss2_mu_libtss2_mu_la_LDFLAGS = -Wl,--version-script=$(srcdir)/lib/tss2-mu.map
 endif # HAVE_LD_VERSION_SCRIPT
-marshal_libmarshal_la_LIBADD = $(libutil)
-marshal_libmarshal_la_SOURCES = $(MARSHAL_SRC)
+src_tss2_mu_libtss2_mu_la_LIBADD = $(libutil)
+src_tss2_mu_libtss2_mu_la_SOURCES = $(TSS2_MU_SRC)
 
 ### TCG TSS TCTI spec libraries ###
 # tcti device library
@@ -121,7 +121,7 @@ tcti_libtcti_device_la_CFLAGS   = $(AM_CFLAGS)
 if HAVE_LD_VERSION_SCRIPT
 tcti_libtcti_device_la_LDFLAGS  = -Wl,--version-script=$(srcdir)/tcti/tcti_device.map
 endif # HAVE_LD_VERSION_SCRIPT
-tcti_libtcti_device_la_LIBADD   = $(libmarshal) $(libutil)
+tcti_libtcti_device_la_LIBADD   = $(libtss2_mu) $(libutil)
 tcti_libtcti_device_la_SOURCES  = tcti/tcti_device.c
 
 # tcti library for microsoft simulator
@@ -135,7 +135,7 @@ tcti_libtcti_mssim_la_CFLAGS   = $(AM_CFLAGS) $(URIPARSER_CFLAGS)
 if HAVE_LD_VERSION_SCRIPT
 tcti_libtcti_mssim_la_LDFLAGS  = -Wl,--version-script=$(srcdir)/lib/tcti-mssim.map
 endif # HAVE_LD_VERSION_SCRIPT
-tcti_libtcti_mssim_la_LIBADD   = $(libmarshal) $(URIPARSER_LIBS) $(libutil)
+tcti_libtcti_mssim_la_LIBADD   = $(libtss2_mu) $(URIPARSER_LIBS) $(libutil)
 tcti_libtcti_mssim_la_SOURCES  = tcti/tcti-mssim.c tcti/sockets.c tcti/sockets.h
 
 ### TCG TSS SAPI spec library ###
@@ -145,8 +145,8 @@ lib_LTLIBRARIES += $(libsapi)
 nodist_pkgconfig_DATA += lib/sapi.pc
 EXTRA_DIST += lib/sapi.pc.in
 
-sysapi_libsapi_la_CFLAGS = $(AM_CFLAGS) -I$(srcdir)/marshal -I$(srcdir)/sysapi/include
-sysapi_libsapi_la_LIBADD  = $(libmarshal)
+sysapi_libsapi_la_CFLAGS = $(AM_CFLAGS) -I$(srcdir)/src/tss2-mu -I$(srcdir)/sysapi/include
+sysapi_libsapi_la_LIBADD  = $(libtss2_mu)
 sysapi_libsapi_la_SOURCES = $(SYSAPI_C) $(SYSAPI_H) $(SYSAPIUTIL_C) \
     $(SYSAPIUTIL_H)
 
@@ -162,7 +162,7 @@ esapi_libesapi_la_CFLAGS  = $(AM_CFLAGS) -Wno-unused-variable -Wno-unused-label 
                             -I$(srcdir)/sysapi/include -I$(srcdir)/esapi/esapi_util \
                             -DESYS_TCTI_DEFAULT_MODULE=$(TCTI_DEFAULT_MODULE) \
                             -DESYS_TCTI_DEFAULT_CONFIG=$(TCTI_DEFAULT_CONFIG)
-esapi_libesapi_la_LIBADD  = $(libsapi) $(libmarshal) $(libtcti_device) $(libtcti_mssim) $(libutil)
+esapi_libesapi_la_LIBADD  = $(libsapi) $(libtss2_mu) $(libtcti_device) $(libtcti_mssim) $(libutil)
 esapi_libesapi_la_LDFLAGS = $(AM_LDFLAGS)  -lgcrypt
 esapi_libesapi_la_SOURCES = $(ESAPI_SRC)
