+++ bootstrap
@@ -28,9 +28,9 @@ echo "Generating file lists: ${VARS_FILE}"
   src_listvar "esapi/esapi_util" "*.h" "ESAPIUTIL_H"
   printf "ESAPI_SRC = \$(ESAPI_H) \$(ESAPI_C) \$(ESAPIUTIL_C) \$(ESAPIUTIL_H)\n"
 
-  src_listvar "marshal" "*.c" "MARSHAL_C"
-  src_listvar "marshal" "*.h" "MARSHAL_H"
-  printf "MARSHAL_SRC = \$(MARSHAL_C) \$(MARSHAL_H)"
+  src_listvar "src/tss2-mu" "*.c" "TSS2_MU_C"
+  src_listvar "src/tss2-mu" "*.h" "TSS2_MU_H"
+  printf "TSS2_MU_SRC = \$(TSS2_MU_C) \$(TSS2_MU_H)"
 ) > ${VARS_FILE}
 
 ${AUTORECONF} --install --sym
