+++ lib/tss2-mu.pc.in
@@ -1,6 +1,6 @@
-Name: sapi
+Name: tss2-mu
 Description: TPM2 type marshalling and unmarshalling library.
 URL: https://github.com/tpm2-software/tpm2-tss
 Version: @VERSION@
 Cflags: -I@includedir@
-Libs: -lmarshal -L@libdir@
+Libs: -ltss2-mu -L@libdir@
