+++ test/tpmclient/tpmclient.int.c
@@ -39,7 +39,7 @@
 #include "../integration/sapi-util.h"
 #include "sample.h"
 #include "tpmclient.h"
-#include "tss2_endian.h"
+#include "tss2-mu/tss2_endian.h"
 #include "sysapi_util.h"
 #include "syscontext.h"
 #define LOGMODULE testtpmclient
